requirements.txt
cn2an==0.5.19
eng_to_ipa==0.0.2
googletrans==3.1.0a0
indic_transliteration==2.3.43
inflect==6.0.2
jamo==0.4.1
jieba==0.42.1
ko_pron==1.3
librosa==0.10.0
matplotlib==3.7.1
nltk==3.7
num_thai==0.0.5
numba==0.56.4
numpy==1.23.5
omegaconf==2.2.3
openai==0.27.1
OpenCC==1.1.1
PyAudio==0.2.13
pyopenjtalk==0.3.0
pypinyin==0.48.0
python-dotenv==1.0.0
scipy==1.10.1
torch==1.13.1
Unidecode==1.3.6
